% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateDensityZIB.R
\name{bivariateDensityZIB}
\alias{bivariateDensityZIB}
\title{Bivariate Zero-Inflated Beta Density}
\usage{
bivariateDensityZIB(x1, p1, mu1, phi1, x2, p2, mu2, phi2, theta, prec = 200)
}
\arguments{
\item{x1}{first vector of relative abundances in the (x1, x2) pair.}

\item{p1}{zero-inflation probability for x1.}

\item{mu1}{beta mean parameter for x1.}

\item{phi1}{beta dispersion parameter for x1.}

\item{x2}{second vector of relative abundances in the (x1, x2) pair.}

\item{p2}{zero-inflation probability for x2.}

\item{mu2}{beta mean parameter for x2.}

\item{phi2}{beta dispersion parameter for x2.}

\item{theta}{copula dependence parameter.}

\item{prec}{the maximal precision to be used in bits.}
}
\value{
\code{bivariateDensityZIB} returns the bivariate zero-inflated beta density function using a frank copula.
}
\description{
Bivariate frank copula density function with zero-inflated beta margins.
}
