% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comicon.R
\name{comicon}
\alias{comicon}
\title{CoMiCoN}
\usage{
comicon(abd, covars = NULL, test = FALSE, ncores = 1)
}
\arguments{
\item{abd}{data frame of microbial relative abundances. Rows are samples and columns are taxa.}

\item{covars}{data frame of covariates to adjust for in the marginal regression models. Rows should be in the same order as \code{abd} and rownames should match.}

\item{test}{logical indicating if two-stage likelihood ratio testing should be performed. Default is \code{FALSE}.}

\item{ncores}{number of cores for parallelization. Default value is 1 (no parallel computing).}
}
\value{
\code{comicon} returns a data frame with two-stage estimated marginal and dependence parameters (and optional two-stage likelihood ratio test statistic) for all possible pairs of microbes.
}
\description{
Implementation of Copulas with Mixture Margins for Covariation Networks using two-stage estimation and testing.
}
