% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dzib.R
\name{dzib}
\alias{dzib}
\title{The Zero-Inflated Beta Distribution}
\usage{
dzib(x, p, mu, phi)
}
\arguments{
\item{x}{vector of quantiles.}

\item{p}{vector of zero-inflation probabilities.}

\item{mu}{vector of means.}

\item{phi}{vector of dispersions.}
}
\value{
\code{dzib} returns the density.
}
\description{
Density function for the zero-inflated beta distribution with zero-inflation probability equal to \code{p}, mean equal to \code{mu} and dispersion equal to \code{phi}.
}
\examples{
dzib(x = 0.8, p = 0.2, mu = 0.5, phi = 0.1)
}
