% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frankConditionalV.R
\name{frankConditionalV}
\alias{frankConditionalV}
\title{The Frank Conditional Copula}
\usage{
frankConditionalV(u, v, theta, prec = 200)
}
\arguments{
\item{u}{vector of uniform random variables.}

\item{v}{vector of uniform random variables.}

\item{theta}{copula dependence parameter.}

\item{prec}{the maximal precision to be used in bits.}
}
\value{
\code{frankConditionalV} returns the frank conditional copula distribution function, conditional on v.
}
\description{
Frank conditional copula function.
}
\examples{
frankConditionalV(u = 0.35, v = 0.5, theta = 2)
}
