% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frankCopula.R
\name{frankCopula}
\alias{frankCopula}
\title{The Frank Copula}
\usage{
frankCopula(u, v, theta, prec = 200)
}
\arguments{
\item{u}{vector of uniform random variables.}

\item{v}{vector of uniform random variables.}

\item{theta}{copula dependence parameter.}

\item{prec}{the maximal precision to be used in bits.}
}
\value{
\code{frankCopula} returns the frank copula distribution function.
}
\description{
Frank copula function.
}
\examples{
frankCopula(u = 0.35, v = 0.5, theta = 2)
}
