% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsLRT.R
\name{lrtTheta}
\alias{lrtTheta}
\title{Two-stage Likelihood Ratio Test}
\usage{
lrtTheta(
  response,
  p1,
  mu1,
  phi1,
  p2,
  mu2,
  phi2,
  theta,
  thetaVar = NULL,
  thetaNull = 0
)
}
\arguments{
\item{response}{data frame with two columns for bivariate pair (x1, x2).}

\item{p1}{two-stage estimate of zero-inflation probability for x1.}

\item{mu1}{two-stage estimate of beta mean parameter for x1.}

\item{phi1}{two-stage estimate of beta dispersion parameter for x1.}

\item{p2}{two-stage estimate of zero-inflation probability for x2.}

\item{mu2}{two-stage estimate of beta mean parameter for x2.}

\item{phi2}{two-stage estimate of beta dispersion parameter for x2.}

\item{theta}{two-stage estimate of copula dependence parameter.}

\item{thetaVar}{jackknife variance of \eqn{\theta}.}

\item{thetaNull}{value of \eqn{\theta} under the null hypothesis. Must be 0 for this implementation of CoMiCoN.}
}
\value{
\code{lrtTheta} returns the two-stage likelihood ratio test statistic.
}
\description{
Two-stage likelihood ratio test for \eqn{\theta}.
}
