% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalMLE.R
\name{marginalMLE}
\alias{marginalMLE}
\title{Marginal Maximum Likelihood Estimation}
\usage{
marginalMLE(formula.mu, formula.phi = ~1, formula.p = ~1, df)
}
\arguments{
\item{formula.mu}{an object of class formula: a symbolic description of the model to be fitted for the beta mean parameter.}

\item{formula.phi}{an object of class formula: a symbolic description of the model to be fitted for the beta dispersion parameter.}

\item{formula.p}{an object of class formula: a symbolic description of the model to be fitted for the zero inflation probability parameter.}

\item{df}{data frame of outcomes and covariates.}
}
\value{
\code{marginalMLE} returns a list of the following elements:
\item{coefficients}{vector of regression coefficients from the models for p, mu, and phi.}
\item{fitted}{data frame of fitted values for p, mu, and phi.}
}
\description{
Maximum likelihood estimation of marginal distribution function parameters.
}
