% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pzib.R
\name{pzib}
\alias{pzib}
\title{The Zero-Inflated Beta Distribution}
\usage{
pzib(x, p, mu, phi)
}
\arguments{
\item{x}{vector of quantiles.}

\item{p}{vector of zero-inflation probabilities.}

\item{mu}{vector of means.}

\item{phi}{vector of dispersions.}
}
\value{
\code{pzib} returns the distribution function.
}
\description{
Distribution function for the zero-inflated beta distribution with zero-inflation probability equal to \code{p}, mean equal to \code{mu} and dispersion equal to \code{phi}.
}
\examples{
pzib(x = 0.8, p = 0.2, mu = 0.5, phi = 0.1)
}
