% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetaTSMLE.R
\name{thetaTSMLE}
\alias{thetaTSMLE}
\title{Two-stage Estimation of \eqn{\theta}}
\usage{
thetaTSMLE(x, lower, upper, p1, mu1, phi1, p2, mu2, phi2, prec = 200)
}
\arguments{
\item{x}{data frame with two columns for bivariate pair (x1, x2).}

\item{lower}{the lower end point of the interval to be searched in univariate optimizer.}

\item{upper}{the lower end point of the interval to be searched in univariate optimizer.}

\item{p1}{two-stage estimate of the zero-inflation probability for x1.}

\item{mu1}{two-stage estimate of the beta mean parameter for x1.}

\item{phi1}{two-stage estimate of the beta dispersion parameter for x1.}

\item{p2}{two-stage estimate of the zero-inflation probability for x2.}

\item{mu2}{two-stage estimate of the beta mean parameter for x2.}

\item{phi2}{two-stage estimate of the beta dispersion parameter for x2.}

\item{prec}{the maximal precision to be used in bits.}
}
\value{
\code{thetaTSMLE} returns the two-stage estimate of \eqn{\theta}.
}
\description{
Two-stage estimation of copula dependence parameter \eqn{\theta}.
}
